Revise the Zab for the finding in the performance experiments

+++ src/main/java/org/vanilladb/comm/protocols/totalorderappl/TotalOrderMessageListener.java
@@ -4,6 +4,6 @@ import java.io.Serializable;
 
 public interface TotalOrderMessageListener {
 	
-    void onRecvTotalOrderMessage(int serialNumber, Serializable message);
+    void onRecvTotalOrderMessage(long serialNumber, Serializable message);
 	
 }
